質問

+++ environment/mytraining/database/migrations/2022_03_04_051824_create_body_part_post_categories_table.php
@@ -14,12 +14,12 @@ class CreateBodypartPostcategoriesTable extends Migration
     public function up()
     {
         Schema::create('body_part_post_categories', function (Blueprint $table) {
-            $table->unsignedInteger('body_part_id');
-            $table->unsignedInteger('post_category_id');
+            $table->unsignedBigInteger('body_part_id');
+            $table->unsignedBigInteger('post_category_id');
             $table->primary(['body_part_id','post_category_id']);
             $table->foreign('body_part_id')->references('id')->on('body_parts');
             $table->foreign('post_category_id')->references('id')->on('post_categories');
-
+            // $table->dropForeign('bodypart_postcategories');
         });
     }
 
@@ -31,5 +31,9 @@ class CreateBodypartPostcategoriesTable extends Migration
     public function down()
     {
         Schema::dropIfExists('bodypart_postcategories');
+        // Schema::table('bodypart_postcategories', function (Blueprint $table) {
+        // $table->foreign('body_part_id')->references('id')->on('body_parts');
+        // $table->foreign('post_category_id')->references('id')->on('post_categories');
+        //  });
     }
 }
